move zmq behaiors to the zmq module
