+++ salt/daemons/flo/zero.py
@@ -12,6 +12,7 @@ import logging
 import hashlib
 import multiprocessing
 import errno
+import time
 # Import ioflo libs
 import ioflo.base.deeding
 # Import third party libs
@@ -250,3 +251,46 @@ class SaltZmqWorker(ioflo.base.deeding.Deed):
             self.created = True
             log.info('Started ZMQ worker')
         self.worker.handle_request()
+
+
+class SaltZmqMaintRotate(ioflo.base.deeding.Deed):
+    '''
+    Update the zmq publish session key
+    '''
+    Ioinits = {'opts': '.salt.opts',
+               'aes': '.salt.var.zmq.aes',
+               'rotate': '.salt.var.zmq.rotate'}
+
+    def action(self):
+        '''
+        Rotate the AES key rotation
+        '''
+        now = time.time()
+        if not self.rotate.value:
+            self.rotate.value = now
+        to_rotate = False
+        dfn = os.path.join(self.opts.value['cachedir'], '.dfn')
+        try:
+            stats = os.stat(dfn)
+            if stats.st_mode == 0o100400:
+                to_rotate = True
+            os.remove(dfn)
+        except os.error:
+            pass
+
+        if self.opts.value.get('publish_session'):
+            if now - self.rotate.value >= self.opts.value['publish_session']:
+                to_rotate = True
+
+        if to_rotate:
+            log.info('Rotating master AES key')
+            # should be unecessary-- since no one else should be modifying
+            with self.aes.value.get_lock():
+                self.aes.value.value = salt.crypt.Crypticle.generate_key_string()
+            #self.event.fire_event({'rotate_aes_key': True}, tag='key')
+            self.rotate.value = now
+            if self.opts.value.get('ping_on_rotate'):
+                # Ping all minions to get them to pick up the new key
+                log.debug('Pinging all connected minions '
+                          'due to AES key rotation')
+                salt.utils.master.ping_all_connected_minions(self.opts.value)
